requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.1.0
fake-useragent==1.1.1
Flask==2.2.2
frozenlist==1.3.3
future==0.18.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.3.0
MarkupSafe==2.1.1
multidict==6.0.4
outcome==1.2.0
PySocks==1.7.1
requests==2.28.1
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.8.2
zipp==3.11.0
